requirements.txt
delegator.py
jinja2
configparser
click
tqdm